pyproject.toml
[tool.poetry]
name = "drf-practice"
version = "0.1.0"
description = ""
authors = ["Hyojin Kwon <raccoonhj33@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django-extensions = "^2.2.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
